Move UDF test to its own test file
